+++ test/bigtest/tests/circulation-rules-test.js
@@ -10,6 +10,7 @@ import { truncate } from 'lodash';
 import setupApplication from '../helpers/setup-application';
 import circulationRules from '../interactors/circulation-rules';
 import { toLowercaseReplaceAllSpaces } from '../helpers/messageConverters';
+import replacer from '../../../src/settings/utils/with-dash-replacer';
 
 const removeDisplayedHints = () => {
   const hints = document.getElementsByClassName('CodeMirror-hints');
@@ -28,13 +29,15 @@ describe('CirculationRules', () => {
   let institutions;
   let campuses;
   let libraries;
+  let locations;
 
   const shortInstitutionCode = 'INST';
-  const longInstitutionCode = 'TESTCODE-TESTCODE-TESTCODE-TESTCODE';
+  const longInstitutionCode = '/TESTCODE TESTCODE TESTCODE TESTCODE';
   const truncatedInstitutionCode = 'TESTCODE-TESTCODE...';
   const institutionsAmount = 30;
   const campusesAmount = 4;
   const librariesAmount = 3;
+  const locationsAmount = 6;
   const truncateOptions = {
     length: 20,
     omission: '...',
@@ -51,15 +54,16 @@ describe('CirculationRules', () => {
     const institutionsWithShortCode = this.server.createList('institution', 1, { code: shortInstitutionCode });
     institutions = await this.server.createList('institution', institutionsAmount - 1, { code: longInstitutionCode });
     institutions.unshift(...institutionsWithShortCode);
-    campuses = await this.server.createList('campus', campusesAmount, { institutionId: institutions[0].id });
-    libraries = this.server.createList('library', librariesAmount, { campusId: campuses[0].id });
+    campuses = await this.server.createList('campus', campusesAmount, { institutionId: institutions[0].id, code: '|campCode >|' });
+    libraries = await this.server.createList('library', librariesAmount, { campusId: campuses[0].id, code: '|libCode *|libCode *' });
+    locations = await this.server.createList('location', locationsAmount, { libraryId: libraries[0].id, code: '=locateCode|', name: 'A-location' });
 
     return this.visit('/settings/circulation/rules', () => {
       expect(circulationRules.$root).to.exist;
     });
   });
 
-  afterEach(function () {
+  afterEach(() => {
     removeDisplayedHints();
   });
 
@@ -76,7 +80,7 @@ describe('CirculationRules', () => {
   });
 
   describe('filter rules', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.filter('term');
     });
 
@@ -86,18 +90,18 @@ describe('CirculationRules', () => {
   });
 
   describe('press Enter without selected hints', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.textArea.focus();
       await circulationRules.editor.pressEnter();
     });
 
     it('should add new line symbol', () => {
-      expect(circulationRules.editor.value).to.be.equal('\n');
+      expect(circulationRules.editor.value).to.equal('\n');
     });
   });
 
   describe('focus the editor', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.textArea.focus();
     });
 
@@ -106,7 +110,7 @@ describe('CirculationRules', () => {
     });
 
     describe('press down arrow key', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.pressArrowDown();
       });
 
@@ -119,7 +123,7 @@ describe('CirculationRules', () => {
       });
 
       describe('press up arrow key', () => {
-        beforeEach(async function () {
+        beforeEach(async () => {
           await circulationRules.editor.pressArrowUp();
         });
 
@@ -129,7 +133,7 @@ describe('CirculationRules', () => {
       });
 
       describe('move selection to the last item', () => {
-        beforeEach(async function () {
+        beforeEach(async () => {
           const hintsCount = circulationRules.editor.hints.sections(0).itemsCount;
           circulationRules.editor.changeActiveItem(hintsCount - 1);
         });
@@ -139,7 +143,7 @@ describe('CirculationRules', () => {
         });
 
         describe('press down arrow key', () => {
-          beforeEach(async function () {
+          beforeEach(async () => {
             await circulationRules.editor.pressArrowDown();
           });
 
@@ -148,7 +152,7 @@ describe('CirculationRules', () => {
           });
 
           describe('press up arrow key', () => {
-            beforeEach(async function () {
+            beforeEach(async () => {
               await circulationRules.editor.pressArrowUp();
             });
 
@@ -162,23 +166,23 @@ describe('CirculationRules', () => {
   });
 
   describe('press Tab without selected hints', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.textArea.focus();
       await circulationRules.editor.pressTab();
     });
 
     it('should add tab symbol', () => {
-      expect(circulationRules.editor.value).to.be.equal('\t');
+      expect(circulationRules.editor.value).to.equal('\t');
     });
   });
 
   describe('entering circulation rule into editor', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.setValue('m book: l example-loan-policy');
     });
 
     it('should set new circulation rule', () => {
-      expect(circulationRules.editor.value).to.be.equal('m book: l example-loan-policy');
+      expect(circulationRules.editor.value).to.equal('m book: l example-loan-policy');
     });
 
     it('is save button enabled after changes', () => {
@@ -187,7 +191,7 @@ describe('CirculationRules', () => {
   });
 
   describe('entering item type', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.setValue('m book: ');
     });
 
@@ -196,16 +200,16 @@ describe('CirculationRules', () => {
     });
 
     it('should display only one hint section', () => {
-      expect(circulationRules.editor.hints.sectionsCount).to.be.equals(1);
+      expect(circulationRules.editor.hints.sectionsCount).to.equal(1);
     });
 
     it('should display policy types menu', () => {
-      expect(circulationRules.editor.hints.sections(0).items(0).text).to.be.equal('Loan policies');
+      expect(circulationRules.editor.hints.sections(0).items(0).text).to.equal('Loan policies');
     });
 
     it('should display the correct header for the policy types list', () => {
       expect(circulationRules.editor.hints.header.isPresent).to.be.true;
-      expect(circulationRules.editor.hints.header.text).to.be.equal('Circulation policies');
+      expect(circulationRules.editor.hints.header.text).to.equal('Circulation policies');
     });
 
     it('should not display the subheader for the policy types list', () => {
@@ -218,7 +222,7 @@ describe('CirculationRules', () => {
   });
 
   describe('choosing non locating type - material type', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.setValue('m ');
     });
 
@@ -231,7 +235,7 @@ describe('CirculationRules', () => {
     });
 
     it('should display only one hint section', () => {
-      expect(circulationRules.editor.hints.sectionsCount).to.be.equals(1);
+      expect(circulationRules.editor.hints.sectionsCount).to.equal(1);
     });
 
     it('should not display subheader', () => {
@@ -244,7 +248,7 @@ describe('CirculationRules', () => {
   });
 
   describe('choosing type locating type - institution', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.setValue('a ');
     });
 
@@ -254,19 +258,19 @@ describe('CirculationRules', () => {
 
     it('should display hints with correct header', () => {
       expect(circulationRules.editor.hints.header.isPresent).to.be.true;
-      expect(circulationRules.editor.hints.header.text).to.be.equal('Select institution');
+      expect(circulationRules.editor.hints.header.text).to.equal('Select institution');
     });
 
     it('should display one hint section', () => {
-      expect(circulationRules.editor.hints.sectionsCount).to.be.equals(1);
+      expect(circulationRules.editor.hints.sectionsCount).to.equal(1);
     });
 
     it('should display hints with correct subheader', () => {
-      expect(circulationRules.editor.hints.sections(0).subheader.text).to.be.equal('Institution');
+      expect(circulationRules.editor.hints.sections(0).subheader.text).to.equal('Institution');
     });
 
     it('should truncate long locations codes', () => {
-      expect(circulationRules.editor.hints.sections(0).items(1).text).to.be.equal(truncatedInstitutionCode);
+      expect(circulationRules.editor.hints.sections(0).items(1).text).to.equal(truncatedInstitutionCode);
     });
 
     it('the section should be scrollable where there are a lot of items', () => {
@@ -274,7 +278,7 @@ describe('CirculationRules', () => {
     });
 
     describe('active the first list item', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(0);
       });
 
@@ -284,7 +288,7 @@ describe('CirculationRules', () => {
     });
 
     describe('select the last item', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(institutionsAmount - 4);
       });
 
@@ -293,7 +297,7 @@ describe('CirculationRules', () => {
       });
 
       describe('select the second item', () => {
-        beforeEach(async function () {
+        beforeEach(async () => {
           await circulationRules.editor.changeActiveItem(1);
         });
 
@@ -304,7 +308,7 @@ describe('CirculationRules', () => {
     });
 
     describe('active item with index, that is greater than the list length, without wrapping avoiding', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(institutionsAmount + 1);
       });
 
@@ -318,7 +322,7 @@ describe('CirculationRules', () => {
     });
 
     describe('active item with index, that is greater than the list length, with wrapping avoiding', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(institutionsAmount + 1, true);
       });
 
@@ -332,7 +336,7 @@ describe('CirculationRules', () => {
     });
 
     describe('active item with negative index without wrapping avoiding', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(-1);
       });
 
@@ -346,7 +350,7 @@ describe('CirculationRules', () => {
     });
 
     describe('active item with negative index with wrapping avoiding', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(-1, true);
       });
 
@@ -360,20 +364,20 @@ describe('CirculationRules', () => {
     });
 
     describe('pick an institution with long name', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(0);
         await circulationRules.editor.pickHint(0);
       });
 
       it('should insert full institution code', () => {
-        expect(circulationRules.editor.value).to.be.equal(
+        expect(circulationRules.editor.value).to.equal(
           `a ${institutions[0].code} `
         );
       });
     });
 
     describe('click on header', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(0);
         await circulationRules.editor.hints.header.click();
       });
@@ -385,7 +389,7 @@ describe('CirculationRules', () => {
     });
 
     describe('click on subheader', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(0);
         await circulationRules.editor.hints.sections(0).subheader.click();
       });
@@ -398,7 +402,7 @@ describe('CirculationRules', () => {
   });
 
   describe('choosing type locating type - campus', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.setValue('b ');
     });
 
@@ -408,45 +412,45 @@ describe('CirculationRules', () => {
 
     it('should display hints with correct header', () => {
       expect(circulationRules.editor.hints.header.isPresent).to.be.true;
-      expect(circulationRules.editor.hints.header.text).to.be.equal('Select campus');
+      expect(circulationRules.editor.hints.header.text).to.equal('Select campus');
     });
 
     it('should display 2 hints sections', () => {
-      expect(circulationRules.editor.hints.sectionsCount).to.be.equals(2);
+      expect(circulationRules.editor.hints.sectionsCount).to.equal(2);
     });
 
     it('should display hints with correct subheaders', () => {
-      expect(circulationRules.editor.hints.sections(0).subheader.text).to.be.equal('Institution');
-      expect(circulationRules.editor.hints.sections(1).subheader.text).to.be.equal('Campus');
+      expect(circulationRules.editor.hints.sections(0).subheader.text).to.equal('Institution');
+      expect(circulationRules.editor.hints.sections(1).subheader.text).to.equal('Campus');
     });
 
     it('the first section should be filled by default', () => {
-      expect(circulationRules.editor.hints.sections(0).items().length).to.be.equal(institutionsAmount);
+      expect(circulationRules.editor.hints.sections(0).items().length).to.equal(institutionsAmount);
     });
 
     it('the second section should be empty by default', () => {
-      expect(circulationRules.editor.hints.sections(1).items().length).to.be.equal(0);
+      expect(circulationRules.editor.hints.sections(1).items().length).to.equal(0);
     });
 
     it('the first section should not contain a special value ANY', () => {
-      expect(circulationRules.editor.hints.sections(0).items(0).text).to.be.equal(institutions[0].code);
+      expect(circulationRules.editor.hints.sections(0).items(0).text).to.equal(institutions[0].code);
     });
 
     describe('pick an institution that contains campuses', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.hints.clickOnItem(0);
       });
 
       it('the second section should be filled by campuses codes', () => {
-        expect(circulationRules.editor.hints.sections(1).items().length).to.be.equal(campusesAmount + 1);
+        expect(circulationRules.editor.hints.sections(1).items().length).to.equal(campusesAmount + 1);
       });
 
       it('the editor value should not be changed', () => {
-        expect(circulationRules.editor.value).to.be.equal('b ');
+        expect(circulationRules.editor.value).to.equal('b ');
       });
 
       it('the second section should contain a special value ANY', () => {
-        expect(circulationRules.editor.hints.sections(1).items(0).text).to.be.equal('<ANY>');
+        expect(circulationRules.editor.hints.sections(1).items(0).text).to.equal('<ANY>');
       });
 
       it('the first item of the second section should be selected by default', () => {
@@ -458,14 +462,14 @@ describe('CirculationRules', () => {
       });
 
       it('the values should be formatted and truncated', () => {
-        const fullValue = `${campuses[0].code} (${institutions[0].code})`;
+        const fullValue = `${replacer(campuses[0].code)} (${institutions[0].code})`;
         const truncatedValue = truncate(fullValue, truncateOptions);
 
-        expect(circulationRules.editor.hints.sections(1).items(1).text).to.be.equals(truncatedValue);
+        expect(circulationRules.editor.hints.sections(1).items(1).text).to.equal(truncatedValue);
       });
 
       describe('move selection in the second section to the last item', () => {
-        beforeEach(async function () {
+        beforeEach(async () => {
           await circulationRules.editor.changeActiveItem(campusesAmount);
         });
 
@@ -478,12 +482,12 @@ describe('CirculationRules', () => {
         });
 
         describe('select selected item in the second section', () => {
-          beforeEach(async function () {
+          beforeEach(async () => {
             await circulationRules.editor.pickHint(campusesAmount);
           });
 
           it('the campus code should be added to the editor value', () => {
-            expect(circulationRules.editor.value).to.be.equal(`b ${campuses[campusesAmount - 1].code} `);
+            expect(circulationRules.editor.value).to.equal(`b ${replacer(campuses[campusesAmount - 1].code)} `);
           });
         });
       });
@@ -491,7 +495,7 @@ describe('CirculationRules', () => {
   });
 
   describe('choosing type locating type - library', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.setValue('c ');
     });
 
@@ -501,48 +505,48 @@ describe('CirculationRules', () => {
 
     it('should display hints with correct header', () => {
       expect(circulationRules.editor.hints.header.isPresent).to.be.true;
-      expect(circulationRules.editor.hints.header.text).to.be.equal('Select library');
+      expect(circulationRules.editor.hints.header.text).to.equal('Select library');
     });
 
     it('should display three hints sections', () => {
-      expect(circulationRules.editor.hints.sectionsCount).to.be.equals(3);
+      expect(circulationRules.editor.hints.sectionsCount).to.equal(3);
     });
 
     it('should display hints with correct subheaders', () => {
-      expect(circulationRules.editor.hints.sections(0).subheader.text).to.be.equal('Institution');
-      expect(circulationRules.editor.hints.sections(1).subheader.text).to.be.equal('Campus');
-      expect(circulationRules.editor.hints.sections(2).subheader.text).to.be.equal('Library');
+      expect(circulationRules.editor.hints.sections(0).subheader.text).to.equal('Institution');
+      expect(circulationRules.editor.hints.sections(1).subheader.text).to.equal('Campus');
+      expect(circulationRules.editor.hints.sections(2).subheader.text).to.equal('Library');
     });
 
     it('the first section should be filled by default', () => {
-      expect(circulationRules.editor.hints.sections(0).items().length).to.be.equal(institutionsAmount);
+      expect(circulationRules.editor.hints.sections(0).items().length).to.equal(institutionsAmount);
     });
 
     it('the second and the third sections should be empty by default', () => {
-      expect(circulationRules.editor.hints.sections(1).items().length).to.be.equal(0);
-      expect(circulationRules.editor.hints.sections(2).items().length).to.be.equal(0);
+      expect(circulationRules.editor.hints.sections(1).items().length).to.equal(0);
+      expect(circulationRules.editor.hints.sections(2).items().length).to.equal(0);
     });
 
     describe('click on the first institution', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.hints.clickOnItem(0);
       });
 
       it('the second section (campus) section should be filled', () => {
-        expect(circulationRules.editor.hints.sections(1).items().length).to.be.equal(campusesAmount + 1);
+        expect(circulationRules.editor.hints.sections(1).items().length).to.equal(campusesAmount + 1);
       });
 
       it('the third section (library) should be empty', () => {
-        expect(circulationRules.editor.hints.sections(2).items().length).to.be.equal(0);
+        expect(circulationRules.editor.hints.sections(2).items().length).to.equal(0);
       });
 
       describe('click on the first campus', () => {
-        beforeEach(async function () {
+        beforeEach(async () => {
           await circulationRules.editor.hints.clickOnItem(1, 1);
         });
 
         it('the third section (library) should be filled', () => {
-          expect(circulationRules.editor.hints.sections(2).items().length).to.be.equal(librariesAmount + 1);
+          expect(circulationRules.editor.hints.sections(2).items().length).to.equal(librariesAmount + 1);
         });
 
         it('the first item of the third section should be selected by default', () => {
@@ -550,19 +554,114 @@ describe('CirculationRules', () => {
         });
 
         it('the values should be formatted and truncated', () => {
-          const fullValue = `${libraries[0].code} (${institutions[0].code}-${campuses[0].code})`;
+          const fullValue = `${replacer(libraries[0].code)} (${institutions[0].code}-${replacer(campuses[0].code)})`;
           const truncatedValue = truncate(fullValue, truncateOptions);
 
-          expect(circulationRules.editor.hints.sections(2).items(1).text).to.be.equals(truncatedValue);
+          expect(circulationRules.editor.hints.sections(2).items(1).text).to.equal(truncatedValue);
         });
 
         describe('click on the first library', () => {
-          beforeEach(async function () {
+          beforeEach(async () => {
             await circulationRules.editor.hints.clickOnItem(1, 2);
           });
 
           it('the library code should be added to the editor value', () => {
-            expect(circulationRules.editor.value).to.be.equal(`c ${libraries[0].code} `);
+            expect(circulationRules.editor.value).to.equal(`c ${replacer(libraries[0].code)} `);
+          });
+        });
+      });
+    });
+  });
+
+  describe('choosing type locating type - location', () => {
+    const rulesEditor = circulationRules.editor;
+    const editorHints = rulesEditor.hints;
+    const getEditorHintSection = sectionIndex => editorHints.sections(sectionIndex);
+
+    beforeEach(async () => {
+      await rulesEditor.setValue('s ');
+    });
+
+    it('should display hints with list to select location codes', () => {
+      expect(editorHints.arePresent).to.be.true;
+    });
+
+    it('should display hints with correct header', () => {
+      expect(editorHints.header.isPresent).to.be.true;
+      expect(editorHints.header.text).to.equal('Select location');
+    });
+
+    it('should display three hints sections', () => {
+      expect(editorHints.sectionsCount).to.equal(4);
+    });
+
+    it('should display hints with correct subheaders', () => {
+      expect(getEditorHintSection(0).subheader.text).to.equal('Institution');
+      expect(getEditorHintSection(1).subheader.text).to.equal('Campus');
+      expect(getEditorHintSection(2).subheader.text).to.equal('Library');
+      expect(getEditorHintSection(3).subheader.text).to.equal('Location');
+    });
+
+    it('the first section should be filled by default', () => {
+      expect(getEditorHintSection(0).items().length).to.equal(institutionsAmount);
+    });
+
+    it('other sections should be empty by default', () => {
+      expect(getEditorHintSection(1).items().length).to.equal(0);
+      expect(getEditorHintSection(2).items().length).to.equal(0);
+      expect(getEditorHintSection(3).items().length).to.equal(0);
+    });
+
+    describe('click on the first institution', () => {
+      beforeEach(async () => {
+        await editorHints.clickOnItem(0);
+      });
+
+      it('the second section (campus) section should be filled', () => {
+        expect(getEditorHintSection(1).items().length).to.equal(campusesAmount + 1);
+      });
+
+      describe('click on the first campus', () => {
+        beforeEach(async () => {
+          await editorHints.clickOnItem(1, 1);
+        });
+
+        it('the third section (library) section should be filled', () => {
+          expect(getEditorHintSection(2).items().length).to.equal(librariesAmount + 1);
+        });
+
+        describe('click on the first library', () => {
+          beforeEach(async () => {
+            await editorHints.clickOnItem(1, 2);
+          });
+
+          it('the fourth section (location) should be filled', () => {
+            expect(getEditorHintSection(3).items().length).to.equal(locationsAmount + 1);
+          });
+
+          it('the fourth section should contain a special value ANY', () => {
+            expect(getEditorHintSection(3).items(0).text).to.equal('<ANY>');
+          });
+
+          it('the first item of the fourth section should be selected by default', () => {
+            expect(getEditorHintSection(3).isFirstItemActive).to.be.true;
+          });
+
+          it('the values should be formatted and truncated', () => {
+            const fullValue = `${replacer(locations[0].code)} (${locations[0].name})`;
+            const truncatedValue = truncate(fullValue, truncateOptions);
+
+            expect(getEditorHintSection(3).items(1).text).to.equal(truncatedValue);
+          });
+
+          describe('click on the first location', () => {
+            beforeEach(async () => {
+              await editorHints.clickOnItem(locationsAmount, 3);
+            });
+
+            it('the location code should be added to the editor value', () => {
+              expect(rulesEditor.value).to.equal(`s ${replacer(locations[locationsAmount - 1].code)} `);
+            });
           });
         });
       });
@@ -570,66 +669,66 @@ describe('CirculationRules', () => {
   });
 
   describe('choosing policy type', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.setValue('m book: ');
     });
 
     describe('press enter on the selected item', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(0);
         await circulationRules.editor.pressEnter();
       });
 
       it('should choose loan policy type', () => {
-        expect(circulationRules.editor.value).to.be.equal('m book: l ');
+        expect(circulationRules.editor.value).to.equal('m book: l ');
       });
     });
 
     describe('press tab on the selected item', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(0);
         await circulationRules.editor.pressTab();
       });
 
       it('should choose loan policy type', () => {
-        expect(circulationRules.editor.value).to.be.equal('m book: l ');
+        expect(circulationRules.editor.value).to.equal('m book: l ');
       });
     });
 
     describe('pick loan policy type', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.changeActiveItem(0);
         await circulationRules.editor.pickHint(0);
       });
 
       it('should choose loan policy type', () => {
-        expect(circulationRules.editor.value).to.be.equal('m book: l ');
+        expect(circulationRules.editor.value).to.equal('m book: l ');
       });
     });
 
     describe('click on first item in policies list', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.hints.clickOnItem(0);
       });
 
       it('should choose loan policy type', () => {
-        expect(circulationRules.editor.value).to.be.equal('m book: l ');
+        expect(circulationRules.editor.value).to.equal('m book: l ');
       });
     });
 
     describe('click on the first item in policies list', () => {
-      beforeEach(async function () {
+      beforeEach(async () => {
         await circulationRules.editor.hints.doubleClickOnItem(0);
       });
 
       it('should choose loan policy type', () => {
-        expect(circulationRules.editor.value).to.be.equal('m book: l ');
+        expect(circulationRules.editor.value).to.equal('m book: l ');
       });
     });
   });
 
   describe('choosing loan policy', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.setValue('m book: ');
       // choose policy type
       await circulationRules.editor.changeActiveItem(0);
@@ -640,14 +739,14 @@ describe('CirculationRules', () => {
     });
 
     it('should choose loan policy', () => {
-      expect(circulationRules.editor.value).to.be.equal(
+      expect(circulationRules.editor.value).to.equal(
         `m book: l ${toLowercaseReplaceAllSpaces(loanPolicies[0].name)} `
       );
     });
   });
 
   describe('choosing fallback policy', () => {
-    beforeEach(async function () {
+    beforeEach(async () => {
       await circulationRules.editor.setValue('fallback-policy: ');
       // choose policy type
       await circulationRules.editor.changeActiveItem(0);
@@ -658,7 +757,7 @@ describe('CirculationRules', () => {
     });
 
     it('should choose loan policy as a fallback', () => {
-      expect(circulationRules.editor.value).to.be.equal(
+      expect(circulationRules.editor.value).to.equal(
         `fallback-policy: l ${toLowercaseReplaceAllSpaces(loanPolicies[0].name)} `
       );
     });
@@ -690,7 +789,7 @@ describe('CirculationRules', () => {
     });
 
     it('should choose loan policy as a fallback', () => {
-      expect(savedRules).to.be.equal(`m 1a54b431-2e4f-452d-9cae-9cee66c9a892 5ee11d91-f7e8-481d-b079-65d708582ccc: l ${lPolicy.id} r ${rPolicy.id} n ${nPolicy.id}`);
+      expect(savedRules).to.equal(`m 1a54b431-2e4f-452d-9cae-9cee66c9a892 5ee11d91-f7e8-481d-b079-65d708582ccc: l ${lPolicy.id} r ${rPolicy.id} n ${nPolicy.id}`);
     });
   });
 });