+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add missing tests for request policy. Part of UICIRC-293.
 * Integrate campus and library menus to circulation rules editor. Refs UICIRC-283.
 * Update `react-to-print` to accept React via `peerDependencies`. UIIN-678.
+* Normalize a non alpha-numeric characters from Location Codes in circulation rules editor. Refs UICIRC-260.
 
 ## [1.9.0](https://github.com/folio-org/ui-circulation/tree/v1.9.0) (2019-07-25)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v1.8.0...v1.9.0)